BM25 = (0.9, 0.4), fixed silly compile-time mistakes
git-svn-id: https://groke.otago.ac.nz/ant/ant@156 7489a7b4-4588-4892-8acd-774bacfab51c

+++ source/memory.h
@@ -49,7 +49,7 @@ if (chunk == NULL)
 	at = chunk = (char *)alloc(&request);
 	allocated = request;
 	if (chunk == NULL)
-		exit(printf("Out of memory:%lld bytes requested\n", bytes));
+		exit(printf("Out of memory:%lld bytes requested\n", (long long)bytes));
 	chunk_end = chunk + allocated;
 	}
 ans = at;
